Update rendering background

+++ src/constants/screen.rs
@@ -1,5 +1,5 @@
-pub const PIXELS_PER_TILE: u8 = 8;
+pub const PIXELS_PER_TILE: u16 = 8;
 pub const PIXEL_SIZE: usize = 3;
-pub const SCREEN_HEIGHT: u8 = 144;
-pub const SCREEN_WIDTH: u8 = 160;
-pub const TILE_WIDTH: u8 = 32;
+pub const SCREEN_HEIGHT: u16 = 144;
+pub const SCREEN_WIDTH: u16 = 160;
+pub const TILE_WIDTH: u16 = 32;
